WIP: Fix the build process, add Makefile
