+++ .gitignore
@@ -5,6 +5,7 @@
 *.so
 *.dylib
 *.rdb
+build/
 
 gibber
