+++ cmd/server/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "gibber/service"
+
+func main() {
+	service.StartServer()
+}
