+++ go.mod
@@ -1,0 +1,20 @@
+module gibber
+
+go 1.13
+
+require (
+	github.com/fatih/structs v1.1.0
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/golang/snappy v0.0.1 // indirect
+	github.com/google/go-cmp v0.3.1 // indirect
+	github.com/mongodb/mongo-go-driver v1.1.3
+	github.com/pkg/errors v0.8.1
+	github.com/stretchr/testify v1.4.0 // indirect
+	github.com/tidwall/pretty v1.0.0 // indirect
+	github.com/xdg/scram v0.0.0-20180814205039-7eeb5667e42c // indirect
+	github.com/xdg/stringprep v1.0.0 // indirect
+	go.mongodb.org/mongo-driver v1.1.3
+	golang.org/x/crypto v0.0.0-20191205180655-e7c4368fe9dd // indirect
+	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e // indirect
+	golang.org/x/text v0.3.2 // indirect
+)
