--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "gibber/server"
-
-func main() {
-	server.StartServer()
-}
